Fixed compiler warning in v-boot reference.
The error was:
  arch/arm/lib/crossystem_arch.c: In function ‘VbReadSharedMemory’:
  arch/arm/lib/crossystem_arch.c:134: error: format ‘%d’ expects type ‘int’, but argument 5 has type ‘long unsigned int’

BUG=none
TEST=(outside choot): cd src/platform/vboot_reference; make

Reviewed-on: http://gerrit.chromium.org/gerrit/1547
Tested-by: Doug Anderson <dianders@chromium.org>

+++ host/arch/arm/lib/crossystem_arch.c
@@ -127,10 +127,10 @@ static int VbReadSharedMemory(void) {
     size = fread(&shared_memory, 1, sizeof(shared_memory), data_file);
     if ((size != sizeof(shared_memory)) || (size != shared_memory.total_size)) {
       fprintf(stderr,  "%s: failed to read shared memory: got %d bytes, "
-              "expected %d, should have been %d\n",
+              "expected %lu, should have been %d\n",
               __FUNCTION__,
               size,
-              sizeof(shared_memory),
+              (unsigned long)sizeof(shared_memory),
               shared_memory.total_size);
       break;
     }
